+++ GIT_UART_OXIMETER_filter/MAX30100.h
@@ -74,8 +74,8 @@ struct configregister { // struct type to represent REGISTERconfig commands rece
 //extern uint8_t MAX_FIFO_DATA[64] ;
 //extern const uint16_t NofSamples;
 extern uint16_t sample_cnt;
-extern uint16_t IR_FIFO_DATA[];
-extern uint16_t RED_FIFO_DATA[];
+extern int32_t IR_FIFO_DATA[];
+extern int32_t RED_FIFO_DATA[];
 extern uint16_t NofSamples;
 
 //*************************************************************************
